Refactor some more
Bench: 9542860
