+++ CMakeLists.txt
@@ -18,7 +18,7 @@ if(LTO_SUPPORTED)
 endif()
 
 
-set(MIDNIGHT_LIB src/engine.cpp src/utils/helpers.cpp src/utils/clock.cpp src/move_search/types.cpp src/move_search/search.cpp src/move_search/reductions.cpp src/move_search/pvs.cpp src/move_search/tables/transposition_table.cpp src/move_search/tables/history_table.cpp src/board/position.cpp src/board/types/bitboard.cpp src/evaluation/pieces/knight.cpp src/evaluation/pieces/pawn.cpp src/evaluation/pieces/queen.cpp src/evaluation/pieces/rook.cpp src/evaluation/pieces/king.cpp src/move_search/tables/lmr_table.cpp src/move_search/move_ordering/move_ordering.cpp src/evaluation/pieces/bishop.cpp src/uci_interpreter/time_manager.cpp)
+set(MIDNIGHT_LIB src/engine.cpp src/utils/helpers.cpp src/utils/clock.cpp src/move_search/types.cpp src/move_search/reductions.cpp src/move_search/pvs.cpp src/move_search/tables/transposition_table.cpp src/move_search/tables/history_table.cpp src/board/position.cpp src/board/types/bitboard.cpp src/evaluation/pieces/knight.cpp src/evaluation/pieces/pawn.cpp src/evaluation/pieces/queen.cpp src/evaluation/pieces/rook.cpp src/evaluation/pieces/king.cpp src/move_search/tables/lmr_table.cpp src/move_search/move_ordering/move_ordering.cpp src/evaluation/pieces/bishop.cpp src/uci_interpreter/time_manager.cpp)
 
 add_executable(Midnight src/main.cpp ${MIDNIGHT_LIB})
 add_executable(MidnightTune texel-tuner/tune_main.cpp ${MIDNIGHT_LIB} texel-tuner/tuner.cpp texel-tuner/threadpool.cpp texel-tuner/midnight.cpp)
