+++ src/move_search/pvs.h
@@ -18,8 +18,9 @@
 extern PVSData data;
 
 void reset_data();
-
-bool position_is_draw(Position &board, const int ply);
+bool position_is_draw(Position &board, int ply);
+void update_best_move_results(int sub_depth, bool debug);
+int scale_soft_time_limit(BestMoveSearchParameters &params, PVSData& ab_results, int depth);
 
 template<Color color>
 int q_search(Position &board, const int ply, int alpha, const int beta) {
@@ -285,4 +286,26 @@ PVSData aspiration_windows(Position& board, int prev_score, short depth, int tim
 	}
 	data.best_move = data.pv.table[0][0];
 	return data;
+}
+
+template<Color color>
+void iterative_deepening(Position& board, BestMoveSearchParameters& params) {
+	reset_clock();
+	std::memset(data.nodes_spend, 0, sizeof(data.nodes_spend));
+	data.nodes_searched = 0;
+	for (int sub_depth = 1; sub_depth <= params.depth; sub_depth++) {
+		int soft_limit = scale_soft_time_limit(params, data, sub_depth);
+		if (time_elapsed_exceeds(soft_limit, TimeResolution::Milliseconds)) {
+			break;
+		}
+		aspiration_windows<color>(board, data.value, sub_depth, params.hard_time_limit);
+		if (data.search_completed) {
+			update_best_move_results(sub_depth, params.debug_info);
+		}
+	}
+}
+
+inline void search(Position& board, BestMoveSearchParameters& parameters) {
+	if (board.turn() == BLACK) iterative_deepening<BLACK>(board, parameters);
+	else iterative_deepening<WHITE>(board, parameters);
 }
